MKOD-987 [FEATURE] Visualise assets download progress + fix app group definition in extension

+++ Assets Bundling POC/Assets Bundling POC/Code/Assets List/AssetsListViewState.swift
@@ -21,7 +21,7 @@ extension AssetListViewRowData {
 
     enum State: Equatable, Hashable {
         case notLoaded
-        case loading(Int)
+        case loading(Double)
         case loaded
         case failed
     }
